- refactor seconds to none config provider into connector namespace - add `aliases` command to list all configured issue key aliases
